32 bit method returns 32 bit

+++ libmorton/include/morton2D_32.h
@@ -35,8 +35,8 @@ inline uint_fast32_t morton2D_32_Encode_magicbits(const uint_fast16_t x, const u
 }
 
 // ENCODE 2D 64-bit morton code: LUT preshifted
-inline uint_fast64_t morton2D_32_Encode_LUT256_shifted(const uint_fast16_t x, const uint_fast16_t y) {
-	uint_fast64_t answer =
+inline uint_fast32_t morton2D_32_Encode_LUT256_shifted(const uint_fast16_t x, const uint_fast16_t y) {
+	uint_fast32_t answer =
 		Morton2D_64_encode_y_256[(y >> 8) & 0x00FF] | // select first 8 bits
 		Morton2D_64_encode_x_256[(x >> 8) & 0x00FF];
 	answer = answer << 16 | // shift by 16 = 2 * 8bits
